Refix example port to proper naming
